Add husky pre-commit hook to lint and test
Currently the test suite runs in a few sections and is pretty quick.
A push to main will auto-deploy so currently keeping checks local
with git hooks instead of a CI environment.

This commit also ignores the prop-type rules for existing components.
In the future this should be done or typed with typescript.